+++ docs/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implement validation for CWE ID chain in a Flaw (OSIDB-357)
 - Implement validation for embargoed flaws not be able to have public trackers (OSIDB-350)
 - Fix Jira tracker created and updated timestamps (OSIDB-14)
+- Fix errata created and updated timestamps (OSIDB-453)
 
 ## [2.3.4] - 2022-12-15
 ### Changed